Single line output functional
