+++ yapr/js/Utils/constants.js
@@ -1,5 +1,14 @@
 /*All the constants that are used by the application*/
 
-const ERRORS = {
-    
+var errors = {
+
+}
+
+const paths = {
+    windows: '/scripts/pystderr.bat',
+    linux: '../../../../scripts/pystderr.sh'
+}
+
+module.exports = {
+    paths, errors
 }
